Prioritize re-authenticating, remove error for missing session middleware, increment version number
Re-authenticating needs to be prioritized because in the event the same user attempts to authenticate the existing session will be used. This is not ideal because a teaching assistant for one course may be a student in another and they would retain their existing role and course, when the expected result is to swap them.

The error was removed because it would cause unexpected behavior for components by throwing an error immediately.